iu_desc_en_lstm/requirements/base.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
Cython==0.29.24
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-Log-Request-ID==0.10.1
fonttools==4.28.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
nltk==3.6.5
numpy==1.22.0
packaging==21.3
Pillow==8.4.0
pycocotools==2.0.2
pytest==6.2.5
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
regex==2021.11.2
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
python-dotenv==0.19.2
